+++ src/main/java/seedu/address/model/person/PersonBuilder.java
@@ -1,20 +1,9 @@
-package seedu.address.testutil;
+package seedu.address.model.person;
 
 import java.util.HashSet;
 import java.util.Set;
 
 import seedu.address.model.appointment.Appointment;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Allergy;
-import seedu.address.model.person.DateOfBirth;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Gender;
-import seedu.address.model.person.MedCon;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Nric;
-import seedu.address.model.person.Person;
-import seedu.address.model.person.Phone;
-import seedu.address.model.person.Priority;
 import seedu.address.model.util.SampleDataUtil;
 
 /**
