+++ src/test/java/seedu/address/logic/commands/AddApptCommandTest.java
@@ -27,7 +27,7 @@ import seedu.address.model.UserPrefs;
 import seedu.address.model.person.Nric;
 import seedu.address.model.person.NricMatchesPredicate;
 import seedu.address.model.person.Person;
-import seedu.address.testutil.PersonBuilder;
+import seedu.address.model.person.PersonBuilder;
 
 /**
  * Contains the integration tests (interaction with the Model) and unit tests for AddApptCommand.
