+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -27,7 +27,7 @@ import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.appointment.Appointment;
 import seedu.address.model.person.NricMatchesPredicate;
 import seedu.address.model.person.Person;
-import seedu.address.testutil.PersonBuilder;
+import seedu.address.model.person.PersonBuilder;
 
 public class AddCommandTest {
