+++ src/test/java/seedu/address/logic/commands/DeleteApptCommandTest.java
@@ -26,7 +26,7 @@ import seedu.address.model.person.AppointmentExistsPredicate;
 import seedu.address.model.person.Nric;
 import seedu.address.model.person.NricMatchesPredicate;
 import seedu.address.model.person.Person;
-import seedu.address.testutil.PersonBuilder;
+import seedu.address.model.person.PersonBuilder;
 
 /**
  * Contains the integration tests (interaction with the Model) and unit tests for DeleteApptCommand.
