+++ src/test/java/seedu/address/model/person/MedConContainsKeywordsPredicateTest.java
@@ -10,9 +10,6 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.testutil.PersonBuilder;
-
-
 
 public class MedConContainsKeywordsPredicateTest {
     public void equals() {
