+++ src/test/java/seedu/address/model/person/NameContainsKeywordsPredicateTest.java
@@ -10,8 +10,6 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.testutil.PersonBuilder;
-
 public class NameContainsKeywordsPredicateTest {
 
     @Test
