+++ src/test/java/seedu/address/model/person/NricMatchesPredicateTest.java
@@ -6,8 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.testutil.PersonBuilder;
-
 public class NricMatchesPredicateTest {
 
     @Test
