+++ src/test/java/seedu/address/model/person/PersonTest.java
@@ -20,8 +20,6 @@ import static seedu.address.testutil.TypicalPersons.BOB;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.testutil.PersonBuilder;
-
 public class PersonTest {
 
     @Test
