+++ src/test/java/seedu/address/model/person/UniquePersonListTest.java
@@ -17,7 +17,6 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.model.person.exceptions.DuplicatePersonException;
 import seedu.address.model.person.exceptions.PersonNotFoundException;
-import seedu.address.testutil.PersonBuilder;
 
 public class UniquePersonListTest {
